separate nakedpromise and xrayvision into utility
